Separate interfaces of Features.Wallet project into their own folder. Bring IWalletFeePolicy under Features.Wallet.Interfaces
