+++ Stratis.Bitcoin.Features.Wallet.Tests/Notifications/BlockObserverTest.cs
@@ -6,6 +6,7 @@ using System.Text;
 using Stratis.Bitcoin.Features.Wallet;
 using Stratis.Bitcoin.Features.Wallet.Notifications;
 using Xunit;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 
 namespace Stratis.Bitcoin.Features.Wallet.Tests.Notifications
 {
