+++ Stratis.Bitcoin.Features.Wallet/Interfaces/IWalletFeePolicy.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Stratis.Bitcoin.Interfaces
+namespace Stratis.Bitcoin.Features.Wallet.Interfaces
 {
     public interface IWalletFeePolicy
     {
