+++ Stratis.Bitcoin.Features.Wallet/Interfaces/IWalletSyncManager.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using NBitcoin;
 
-namespace Stratis.Bitcoin.Features.Wallet
+namespace Stratis.Bitcoin.Features.Wallet.Interfaces
 {
 	public interface IWalletSyncManager
 	{
