+++ Stratis.Bitcoin.IntegrationTests/NodeBuilder.cs
@@ -12,6 +12,7 @@ using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.Features.Miner;
 using Stratis.Bitcoin.Features.RPC;
 using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
