+++ Stratis.Bitcoin.IntegrationTests/WalletTests.cs
@@ -2,6 +2,7 @@
 using NBitcoin;
 using Stratis.Bitcoin.Features.Consensus;
 using Stratis.Bitcoin.Features.Wallet;
+using Stratis.Bitcoin.Features.Wallet.Interfaces;
 using System.Collections.Generic;
 using System.Linq;
 using Xunit;
